通知の表示設定を移動
